Improve the speed of the domain generator. On the cpu path, move to FFTW when available. On the gpu path, re-write some kernels to be more efficient (in particular, use fewer memory transfers). Change the way the bias is calculated from x-x**3 to tanh(x)

+++ speckle/kernels/domains_mkavailable.cl
@@ -1,10 +1,10 @@
 __kernel void execute(
-    __global uchar* available,
+    __global float* available,
     __global uchar* pospins,
     __global uchar* negpins,
     __global uchar* walls)
 
     {	
 	int i = get_global_id(0);
-	available[i] = (uchar)(walls[i]*pospins[i]*negpins[i]);
+	available[i] = (float)(walls[i]*pospins[i]*negpins[i]);
     }
